CSV hex output & regex fix for odd case.
Done for the day. This project will take some thinking, and almost definitely some AI interaction, to analyse the pulses properly.

At the moment all we can output is into a nicely formatted CSV to manually look for patterns (i.e. headers, checksums) but that is going to take me a fair amount of time and testing signals. I reckon AI could point out these patterns 100x quicker than myself so that is the next step.